requirements.txt
flask
python-dotenv
openai
groq
torch
transformers
datasets
seqeval
pandas
numpy
scikit-learn
openai-whisper
google-cloud-speech
google-cloud-translate
reportlab
gunicorn
